requirements.txt
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
pydub==0.25.1
python-magic-bin==0.4.14 